+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 0.6.8 (Unreleased)
 * Teach scss-lint that `border: 0;` is preferred over `border: none;`
+* Teach scss-lint that variable names should not contain underscores or capital
+  letters
 
 0.6.7
 * Fixed `--ignore-linters` flag